+++ src/main/java/sensordata/sadd/pages/LoginPage.form
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="main.java.sensordata.sadd.LoginPage">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="main.java.sensordata.sadd.pages.LoginPage">
   <grid id="27dc6" binding="loginPanel" layout-manager="GridLayoutManager" row-count="5" column-count="12" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
     <margin top="0" left="0" bottom="0" right="0"/>
     <constraints>
