+++ src/main/java/sensordata/sadd/pages/NewPassword.java
@@ -1,6 +1,6 @@
-package main.java.sensordata.sadd;
+package main.java.sensordata.sadd.pages;
 
-import main.java.sensordata.sadd.pages.Page;
+import main.java.sensordata.sadd.database.UserInfo;
 
 import javax.swing.*;
 import java.awt.*;
