+++ src/main/java/sensordata/sadd/pages/ResetPage.java
@@ -1,6 +1,7 @@
-package main.java.sensordata.sadd;
+package main.java.sensordata.sadd.pages;
 
-import main.java.sensordata.sadd.pages.Page;
+import main.java.sensordata.sadd.MailUtil;
+import main.java.sensordata.sadd.Main;
 
 import javax.swing.*;
 import java.awt.*;
@@ -45,10 +46,10 @@ public class ResetPage extends Page implements ActionListener {
                 break;
             case REQUEST:
                 email = textField1.getText();
-                JavaMailUtil random_generator = new JavaMailUtil();
+                MailUtil random_generator = new MailUtil();
                 number = random_generator.random_numbers();
                 try {
-                    JavaMailUtil.sendMail(email, number);
+                    MailUtil.sendMail(email, number);
                 } catch (Exception E) {
                 }
                 break;
